db/requirements.txt
pipautopep8==2.0.2
bpyutils==0.5.8
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
six==1.16.0
toml==0.10.2
Werkzeug==2.2.3